Refactor time calculation to utilities
