+++ utilities/index.js
@@ -1,3 +1,4 @@
 export * from './weatherIcon';
+export * from './TimeCalculator';
 
 export default {};
